Rename composable folder and add a new one for each step
